re-organize, documentation
